backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
google-generativeai==0.8.3
requests==2.32.5
pillow==11.3.0
pydantic==2.11.7
gtts==2.5.1


